requirements.txt
absl-py==1.4.0
accelerate==0.18.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dlib @ file:///D:/FYP%20Project/ocr/dlib-19.22.99-cp39-cp39-win_amd64.whl
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.10.7
Flask==2.2.3
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
flatbuffers==23.3.3
fonttools==4.39.2
gast==0.4.0
google-api-core==2.11.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-vision==3.4.1
google-pasta==0.2.0
googleapis-common-protos==1.59.0
GPUtil==1.4.0
grpcio==1.53.0
grpcio-status==1.53.0
h5py==3.8.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
imutils==0.5.4
ipython==8.11.0
itsdangerous==2.1.2
jax==0.4.7
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
labelImg==1.8.6
lazy_loader==0.2
Levenshtein==0.20.9
libclang==16.0.0
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.0.4
mpmath==1.3.0
mysqlclient==2.1.1
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==2.15.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.18.0
six==1.16.0
sklearn==0.0.post1
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.21
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
ultralytics==8.0.58
urllib3==1.26.15
utils==1.0.1
wcwidth==0.2.6
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.14.1
zipp==3.15.0
